update
trying to make things more legible
